Entity layer created closes #3

+++ src/main/java/com/clubmaster/clubmaster/entity/Team.java
@@ -12,7 +12,7 @@ public class Team {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(unique = true, nullable = false, length = 100)
     private String name;
